add callbacks for refreshing state of previous pages on Navigator.pop
This is a bit ugly, but seems necessary in order to trigger redrawing
with new state upon Navigator.pop. The popped page might have modified
global state, so the previous page must be refreshed.

Instead of redrawing everything, the alternative would be to track
changes more precisely within the internal app state, and attempt to
overwrite previous global state, and redraw.