defer message about unknown channels until a package is actually not found
This allows to work with local copies of channels (i.e. the remote
channel URL is not in your channels) without immediate error message
when opening a package using sc4pac:// scheme for a remote URL – as long
as all packages are found. Even if some packages are not found, opening
the others is still possible.