add dialog for adding new channel when a package is not found
When one or more packages are opened using a sc4pac:// URL, but aren't
found in any configured channel, the channels passed through the URL are
added to the profile (after user confirmation).

Additionally, this improves error handling for the `channelStats`
Future, e.g. when a URL is malformed or does not point to a valid
channel, a alert dialog will be displayed.